requirements.txt
anthropic>=0.49.0
docker>=7.1.0
instructor[claude]>=1.7.3
loguru>=0.7.3
mcp[cli]>=1.6.0
pydantic>=2.10.6
python-dotenv>=1.0.1
rich>=13.9.4
couchdb
asyncpg
pyyaml
jsonschema
psycopg2-binary
fastmcp
boto3

# Development dependencies
types-docker>=7.1.0.20241229
# types-pydantic>=2.10.6
# types-rich>=13.9.4
# types-python-dotenv>=1.0.1
# types-loguru>=0.7.3
# types-instructor>=1.7.3